feat: implement user registration and login functionality with validation and file upload

+++ src/utils/cloudinary.js
@@ -1,25 +1,30 @@
-import {v2 as cloudinary} from "cloudinary"
-import fs from "fs"
+import { v2 as cloudinary } from 'cloudinary';
+import fs from 'fs';
 
 cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET,
-})
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
 
 const uploadOnCloudinary = async (localFilePath) => {
-    try {
-        if(!localFilePath) return null
-        const file = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
-        console.log("file uploaded: ", file.url);
-        return url
-        
-    } catch (error) {
-        fs.unlinkSync(localFilePath);
+  try {
+    if (!localFilePath) return null;
+    const file = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: 'auto',
+    });
+    console.log('file uploaded: ', file.url);
+    // fs.unlinkSync(localFilePath); // for future when the cloudinay successfully upload file on server 
+    return file;
+  } catch (error) {
+    console.error('Cloudinary upload error:', error);
+
+    // delete file if exists
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
-}
+    return null;
+  }
+};
 
-
-export {uploadOnCloudinary}
+export { uploadOnCloudinary };
